requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
bs4==0.0.1
click==8.1.3
Deprecated==1.2.13
docopt==0.6.2
fastapi==0.88.0
h11==0.14.0
hangul-romanize==0.1.0
idna==3.4
install==1.3.5
ipadic==1.0.0
jaconv==0.3
jieba==0.42.1
joblib==1.2.0
mecab-python3==1.0.5
numpy==1.23.5
pinyin==0.4.0
pydantic==1.10.2
pykakasi==2.2.1
python-crfsuite==0.9.8
python-random-strings==0.3
pyvi==0.1.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tabulate==0.9.0
typing_extensions==4.4.0
unidic-lite==1.0.8
uvicorn==0.20.0
webvtt-py==0.4.6
wrapt==1.14.1
